Refactor: add getPercentage function and combine devise.js to global.js
